requirements.txt
gcp-cli
ncclient
netaddr
paramiko
hvac
PyNaCl
pynacl
requests
